🚚 move main logic of marketplace_core into lib.rs
