+++ marketplace-core/src/routes/applications/list/mod.rs
@@ -1,6 +1,6 @@
+use crate::dto;
 use http_api_problem::HttpApiProblem;
 use itertools::Itertools;
-use marketplace_core::dto;
 use marketplace_domain::ApplicationProjectionRepository;
 use rocket::{serde::json::Json, State};
 use rocket_okapi::openapi;
