+++ marketplace-core/src/routes/contributions/applications_list/mod.rs
@@ -1,8 +1,8 @@
 use std::sync::Arc;
 
+use crate::dto;
 use http_api_problem::HttpApiProblem;
 use itertools::Itertools;
-use marketplace_core::dto;
 use marketplace_domain::{ApplicationProjection, ApplicationProjectionRepository, ContributorId};
 use rocket::{serde::json::Json, State};
 use rocket_okapi::openapi;
