+++ marketplace-core/src/routes/projects/refresh.rs
@@ -1,10 +1,10 @@
 use crate::{
+	application::{RefreshLeadContributors, RefreshProjectsMembers},
 	routes::{api_key::ApiKey, to_http_api_problem::ToHttpApiProblem},
 	RefreshProjects,
 };
 use futures::try_join;
 use http_api_problem::HttpApiProblem;
-use marketplace_core::application::{RefreshLeadContributors, RefreshProjectsMembers};
 use rocket::{http::Status, State};
 use rocket_okapi::openapi;
